+++ packages/matter.js/src/cluster/client/ClusterClientTypes.ts
@@ -11,7 +11,7 @@ import {
 import { AttributeClient } from "./AttributeClient.js";
 import { Merge } from "../../util/Type.js";
 import { InteractionClient } from "../../protocol/interaction/InteractionClient.js";
-import { ClusterServerObj } from "../server/ClusterServer.js";
+import { ClusterServerObj } from "../server/ClusterServerTypes.js";
 import { EventClient } from "./EventClient.js";
 import { BitSchema } from "../../schema/BitmapSchema.js";
 import { DecodedEventData } from "../../protocol/interaction/EventDataDecoder.js";
