+++ packages/matter.js/src/cluster/server/GroupKeyManagementServer.ts
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { ClusterServerHandlers } from "./ClusterServer.js";
+import { ClusterServerHandlers } from "./ClusterServerTypes.js";
 import { GroupKeyManagementCluster } from "../definitions/GroupKeyManagementCluster.js";
 import { NotImplementedError } from "../../common/MatterError.js";
