+++ packages/matter.js/src/device/Aggregator.ts
@@ -6,7 +6,7 @@
 import { Device } from "./Device.js";
 import { DeviceTypes } from "./DeviceTypes.js";
 import { ComposedDevice } from "./ComposedDevice.js";
-import { AttributeInitialValues } from "../cluster/server/ClusterServer.js";
+import { AttributeInitialValues } from "../cluster/server/ClusterServerTypes.js";
 import { BridgedDeviceBasicInformationCluster } from "../cluster/definitions/BridgedDeviceBasicInformationCluster.js";
 import { ClusterServer } from "../protocol/interaction/InteractionServer.js";
 import { Endpoint, EndpointOptions } from "./Endpoint.js";
