+++ packages/matter.js/src/device/Device.ts
@@ -8,8 +8,8 @@ import { DeviceClasses, DeviceTypeDefinition, DeviceTypes } from "./DeviceTypes.
 import { Endpoint, EndpointOptions } from "./Endpoint.js";
 import { AtLeastOne } from "../util/Array.js";
 import { HandlerFunction, NamedHandler } from "../util/NamedHandler.js";
-import { ClusterClientObj, isClusterClient } from "../cluster/client/ClusterClient.js";
-import { ClusterServerObj, isClusterServer } from "../cluster/server/ClusterServer.js";
+import { ClusterClientObj, isClusterClient } from "../cluster/client/ClusterClientTypes.js";
+import { ClusterServerObj, isClusterServer } from "../cluster/server/ClusterServerTypes.js";
 import { Attributes, Cluster, Commands, Events } from "../cluster/Cluster.js";
 import { BitSchema, TypeFromPartialBitSchema } from "../schema/BitmapSchema.js";
 import { BindingCluster } from "../cluster/definitions/BindingCluster.js";
