+++ packages/matter.js/src/device/Endpoint.ts
@@ -13,8 +13,8 @@ import { Attributes, Cluster, Commands, Events } from "../cluster/Cluster.js";
 import { EndpointNumber } from "../datatype/EndpointNumber.js";
 import { FixedLabelCluster } from "../cluster/definitions/FixedLabelCluster.js";
 import { UserLabelCluster } from "../cluster/definitions/UserLabelCluster.js";
-import { ClusterClientObj } from "../cluster/client/ClusterClient.js";
-import { ClusterServerObj, ClusterServerObjForCluster } from "../cluster/server/ClusterServer.js";
+import { ClusterClientObj } from "../cluster/client/ClusterClientTypes.js";
+import { ClusterServerObj, ClusterServerObjForCluster } from "../cluster/server/ClusterServerTypes.js";
 import { InteractionClient } from "../protocol/interaction/InteractionClient.js";
 import { BasicInformationCluster } from "../cluster/definitions/BasicInformationCluster.js";
 import { BridgedDeviceBasicInformationCluster } from "../cluster/definitions/BridgedDeviceBasicInformationCluster.js";
